+++ packages/gatsby-codemods/src/transforms/__testfixtures__/global-graphql-calls/no-import-esm.input.js
@@ -6,4 +6,4 @@ export const query = graphql`
       prefix
     }
   }
-`
+`;