+++ packages/gatsby-plugin-offline/src/gatsby-node.js
@@ -3,6 +3,7 @@ const precache = require(`sw-precache`)
 const path = require(`path`)
 const slash = require(`slash`)
 const _ = require(`lodash`)
+const replace = require(`replace-in-file`)
 
 const getResourcesFromHTML = require(`./get-resources-from-html`)
 
@@ -46,8 +47,14 @@ exports.onPostBuild = (args, pluginOptions) => {
     rootDir
   )
 
-  const criticalFilePaths = getResourcesFromHTML(
-    `${process.cwd()}/${rootDir}/index.html`
+  const criticalFilePaths = _.uniq(
+    _.concat(
+      getResourcesFromHTML(`${process.cwd()}/${rootDir}/index.html`),
+      getResourcesFromHTML(`${process.cwd()}/${rootDir}/404.html`),
+      getResourcesFromHTML(
+        `${process.cwd()}/${rootDir}/offline-plugin-app-shell-fallback/index.html`
+      )
+    )
   )
 
   const options = {
@@ -65,21 +72,22 @@ exports.onPostBuild = (args, pluginOptions) => {
     // https://github.com/GoogleChrome/sw-precache#replaceprefix-string
     replacePrefix: args.pathPrefix || ``,
     navigateFallback: `/offline-plugin-app-shell-fallback/index.html`,
-    // Only match URLs without extensions.
+    // Only match URLs without extensions or the query `no-cache=1`.
     // So example.com/about/ will pass but
+    // example.com/about/?no-cache=1 and
     // example.com/cheeseburger.jpg will not.
     // We only want the service worker to handle our "clean"
     // URLs and not any files hosted on the site.
     //
-    // Regex from http://stackoverflow.com/a/18017805
-    navigateFallbackWhitelist: [/^.*([^.]{5}|.html)$/],
+    // Regex based on http://stackoverflow.com/a/18017805
+    navigateFallbackWhitelist: [/^.*([^.]{5}|.html)(?<!(\?|&)no-cache=1)$/],
     cacheId: `gatsby-plugin-offline`,
     // Don't cache-bust JS files and anything in the static directory
     dontCacheBustUrlsMatching: /(.*js$|\/static\/)/,
     runtimeCaching: [
       {
-        // Add runtime caching of images.
-        urlPattern: /\.(?:png|jpg|jpeg|webp|svg|gif|tiff|js|woff|woff2)$/,
+        // Add runtime caching of various page resources.
+        urlPattern: /\.(?:png|jpg|jpeg|webp|svg|gif|tiff|js|woff|woff2|json|css)$/,
         handler: `fastest`,
       },
     ],
@@ -88,5 +96,19 @@ exports.onPostBuild = (args, pluginOptions) => {
 
   const combinedOptions = _.defaults(pluginOptions, options)
 
-  return precache.write(`public/sw.js`, combinedOptions)
+  return precache.write(`public/sw.js`, combinedOptions).then(() =>
+    // Patch sw.js to include search queries when matching URLs against navigateFallbackWhitelist
+    replace({
+      files: `public/sw.js`,
+      from: `path = (new URL(absoluteUrlString)).pathname`,
+      to: `url = new URL(absoluteUrlString), path = url.pathname + url.search`,
+    }).then(changes => {
+      // Check that the patch has been applied correctly
+      if (changes.length !== 1)
+        throw new Error(
+          `Patching sw.js failed - sw-precache has probably been modified upstream.\n` +
+            `Please report this issue at https://github.com/gatsbyjs/gatsby/issues`
+        )
+    })
+  )
 }