+++ packages/gatsby-plugin-styletron/src/gatsby-browser.js
@@ -1,6 +1,8 @@
-import { createElement as h } from "react"
-import styletron from "./index.js"
-import { Provider } from "styletron-react"
+const React = require(`react`)
+const styletron = require(`./index`)
+const { Provider } = require(`styletron-react`)
 
-exports.wrapRootComponent = ({ Root }, options) => () =>
-  h(Provider, { value: styletron(options).instance }, h(Root))
+// eslint-disable-next-line react/prop-types
+exports.wrapRootElement = ({ element }, options) => (
+  <Provider value={styletron(options).instance}>{element}</Provider>
+)