+++ packages/gatsby-transformer-javascript-static-exports/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.1-rc.0"></a>
+
+## [2.1.1-rc.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-javascript-static-exports/compare/gatsby-transformer-javascript-static-exports@2.1.1-beta.3...gatsby-transformer-javascript-static-exports@2.1.1-rc.0) (2018-08-21)
+
+**Note:** Version bump only for package gatsby-transformer-javascript-static-exports
+
 <a name="2.1.1-beta.3"></a>
 
 ## [2.1.1-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-javascript-static-exports/compare/gatsby-transformer-javascript-static-exports@2.1.1-beta.2...gatsby-transformer-javascript-static-exports@2.1.1-beta.3) (2018-07-21)